Buttons work for closet when navigating
